登録と編集画面のenumを共通化
