+++ RandomChoiceApp/Model/eeeenum.swift
@@ -1,0 +1,9 @@
+//
+//  eeeenum.swift
+//  RandomChoiceApp
+//
+//  Created by Kazuma Fuchi on 2020/12/17.
+//  Copyright © 2020 AYANO HARA. All rights reserved.
+//
+
+import Foundation
